refactor: split SubredditRow helpers into extension file to stay under 300 lines
Move computed properties (effectivePeakDays, effectivePeakHours,
showsSuggested, etc.), action methods (toggleHourLocal, applyPreset),
and SubredditDropDelegate into SubredditRowHelpers.swift.
