+++ Sources/Views/SubredditRow.swift
@@ -30,7 +30,7 @@ struct SubredditRow: View {
               .accessibilityIdentifier("channels.subredditRow.\(sub.id.uuidString).remove")
           } else {
             VStack(alignment: .trailing, spacing: 2) {
-              Text(peakDaysSummary)
+              Text(peakDaysSummaryText)
                 .font(.system(size: 10))
                 .foregroundStyle(.secondary)
               if hasPostingChecklist {
@@ -221,11 +221,6 @@ struct SubredditRow: View {
     }
   }
 
-  private func applyPreset(_ preset: SubredditPeakSelection.PeakPreset) {
-    let applied = SubredditPeakSelection.applyPreset(preset)
-    sub.peakDaysOverride = applied.days
-    sub.peakHoursUtcOverride = applied.utcHours
-  }
 
   private var eventSourceChips: some View {
     HStack(spacing: 6) {
@@ -249,102 +244,8 @@ struct SubredditRow: View {
       .foregroundStyle(count > 0 ? color : .secondary)
   }
 
-  private func toggleHourLocal(_ localHour: Int) {
-    let utcHour = SubredditPeakSelection.localHourToUtc(localHour)
-    if showsSuggested {
-      let suggested = SubredditPeakSelection.suggestedDefaults()
-      sub.peakDaysOverride = suggested.days
-      sub.peakHoursUtcOverride = SubredditPeakSelection.toggledHour(utcHour, in: suggested.utcHours)
-    } else {
-      sub.peakHoursUtcOverride = SubredditPeakSelection.toggledHour(utcHour, in: sub.peakHoursUtcOverride)
-    }
-  }
-
-  private var peakDaysSummary: String {
-    SubredditPeakSelection.peakDaysSummary(
-      effectivePeakDays: effectivePeakDays, hasOverride: hasOverride)
-  }
-
   private func resetDefaults() {
     sub.peakDaysOverride = nil
     sub.peakHoursUtcOverride = nil
-  }
-
-  private var hasOverride: Bool {
-    SubredditPeakSelection.hasOverride(days: sub.peakDaysOverride, hours: sub.peakHoursUtcOverride)
-  }
-
-  private var showsSuggested: Bool {
-    SubredditPeakSelection.needsSuggestedDefaults(
-      daysOverride: sub.peakDaysOverride,
-      hoursOverride: sub.peakHoursUtcOverride,
-      peakInfo: peakInfo
-    )
-  }
-
-  private var effectivePeakDays: [String] {
-    if showsSuggested {
-      return SubredditPeakSelection.suggestedDefaults().days
-    }
-    return SubredditPeakSelection.effectivePeakDays(override: sub.peakDaysOverride, peakInfo: peakInfo)
-  }
-
-  private var effectivePeakHours: [Int] {
-    if showsSuggested {
-      return SubredditPeakSelection.suggestedDefaults().utcHours
-    }
-    return SubredditPeakSelection.effectivePeakHours(
-      override: sub.peakHoursUtcOverride, peakInfo: peakInfo)
-  }
-
-  private var effectivePeakHoursLocal: [Int] {
-    if showsSuggested {
-      return SubredditPeakSelection.suggestedDefaults().localHours
-    }
-    return SubredditPeakSelection.utcHoursToLocal(effectivePeakHours)
-  }
-
-  private var eventSourceSummary: EventSourceSummary {
-    EventSourceSummary.active(events: sub.events)
-  }
-
-  private var hasPostingChecklist: Bool {
-    !(sub.postingChecklist?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty ?? true)
-  }
-
-  private var postingChecklistBinding: Binding<String> {
-    Binding(
-      get: { sub.postingChecklist ?? "" },
-      set: { newValue in
-        let trimmed = newValue.trimmingCharacters(in: .whitespacesAndNewlines)
-        sub.postingChecklist = trimmed.isEmpty ? nil : newValue
-      }
-    )
-  }
-}
-
-struct SubredditDropDelegate: DropDelegate {
-  let target: Subreddit
-  @Binding var dragging: Subreddit?
-  let subreddits: [Subreddit]
-  let modelContext: ModelContext
-
-  func performDrop(info: DropInfo) -> Bool {
-    dragging = nil
-    return true
-  }
-
-  func dropEntered(info: DropInfo) {
-    guard let source = dragging, source.id != target.id else { return }
-    SubredditPersistenceActions.reorder(
-      source: source,
-      target: target,
-      subreddits: subreddits,
-      modelContext: modelContext
-    )
-  }
-
-  func dropUpdated(info: DropInfo) -> DropProposal? {
-    DropProposal(operation: .move)
   }
 }
